Remove ExecutionState model
